+++ src/index.ts
@@ -9,3 +9,4 @@ export * from "./annotations.js";
 export * from "./logger.js";
 export * from "./errors.js";
 export * from "./cache-status.js";
+export * from "./target-resolver.js";